[Added] - Mesh class that handles the loading and rendering of individual of meshes
I did find that the model loading does not work correctly on specific models, like the Chess board. I need to figure out what's causing that to happen.
