+++ LAGGED/src/Platform/OpenGL/Renderer/GL_Model.cpp
@@ -73,7 +73,12 @@ namespace LAG
 			WARNING("Warning while loading Model: {0}", warningMsg);
 
 		std::string modelDirPath = std::filesystem::path(filePath).parent_path().string();
-		LoadModel(*m_Model);
+		
+		m_Meshes.resize(m_Model->meshes.size());
+		for (int i = 0; i < m_Model->meshes.size(); i++)
+		{
+			m_Meshes[i].Load(GetPath().GetString(), *m_Model, i);
+		}
 
 		m_PreTransformScale = 1.f;
 		return true;
@@ -81,128 +86,9 @@ namespace LAG
 
 	bool Model::Unload()
 	{
-		m_Buffer.Shutdown();
+		for (auto& it : m_Meshes)
+			it.Unload();
 		return true;
-	}
-
-	std::vector<MeshData> LoadVertices(tinygltf::Model& modelData, tinygltf::Primitive& primitive)
-	{
-		std::vector<MeshData> meshData;
-
-		//Load vertex data
-		{
-			const auto& primitiveAttributes = primitive.attributes["POSITION"];
-			const auto& accessors = modelData.accessors[primitiveAttributes];
-			const auto& bufferViews = modelData.bufferViews[accessors.bufferView];
-			const auto& buffers = modelData.buffers[bufferViews.buffer];
-
-			meshData.reserve(accessors.count);
-			const float* positions = reinterpret_cast<const float*>(&buffers.data[bufferViews.byteOffset + accessors.byteOffset]);
-			for (size_t i = 0; i < meshData.capacity(); i++)
-			{
-				meshData.emplace_back(MeshData({}));
-				meshData[i].vertices = glm::vec3(positions[i * 3 + 0], positions[i * 3 + 1], positions[i * 3 + 2]);
-			}
-		}
-
-		//Load normals
-		{
-			const auto& primitiveAttributes = primitive.attributes["NORMAL"];
-			const auto& accessors = modelData.accessors[primitiveAttributes];
-			const auto& bufferViews = modelData.bufferViews[accessors.bufferView];
-			const auto& buffers = modelData.buffers[bufferViews.buffer];
-			const float* normals = reinterpret_cast<const float*>(&buffers.data[bufferViews.byteOffset + accessors.byteOffset]);
-			for (size_t i = 0; i < meshData.capacity(); i++)
-			{
-				meshData[i].normals = glm::vec3(normals[i * 3 + 0], normals[i * 3 + 1], normals[i * 3 + 2]);
-			}
-		}
-
-		//Load texture coordinates
-		{
-			const auto& primitiveAttributes = primitive.attributes["TEXCOORD_0"];
-			const auto& accessors = modelData.accessors[primitiveAttributes];
-			const auto& bufferViews = modelData.bufferViews[accessors.bufferView];
-			const auto& buffers = modelData.buffers[bufferViews.buffer];
-
-			if (accessors.componentType != TINYGLTF_COMPONENT_TYPE_FLOAT)
-			{
-				ERROR("Incorrect component type detected while loading mesh.");
-			}
-
-			const float* positions = reinterpret_cast<const float*>(&buffers.data[bufferViews.byteOffset + accessors.byteOffset]);
-			for (size_t i = 0; i < meshData.capacity(); i++)
-			{
-				meshData[i].textureCoords = glm::vec2(positions[i * 2 + 0], positions[i * 2 + 1]);
-			}
-		}
-
-		return meshData;
-	}
-
-	std::vector<unsigned int> LoadIndices(tinygltf::Model& modelData, tinygltf::Primitive& primitive)
-	{
-		std::vector<unsigned int> indices;
-
-		const auto& accessors = modelData.accessors[primitive.indices];
-		const auto& bufferViews = modelData.bufferViews[accessors.bufferView];
-		const auto& buffers = modelData.buffers[bufferViews.buffer];
-		size_t accessorSize = tinygltf::GetNumComponentsInType(accessors.type) * tinygltf::GetComponentSizeInBytes(accessors.componentType);
-		indices.reserve((bufferViews.byteLength / accessorSize));
-		{
-			auto* indexArray = &buffers.data.data()[bufferViews.byteOffset];
-			const unsigned short* indexArrayShort = reinterpret_cast<const unsigned short*>(indexArray);
-
-			for (size_t i = 0; i < indices.capacity(); i++)
-			{
-				indices.emplace_back(indexArrayShort[i]);
-			}
-		}
-
-		return indices;
-	}
-
-	std::vector<size_t> LoadTexture(tinygltf::Model& modelData, std::string modelPath)
-	{
-		std::vector<size_t> textures;
-
-		tinygltf::Texture& texture = modelData.textures.at(0);
-		std::string textureName = modelData.images[texture.source].uri; //Get texture resource identifier
-
-		std::string texturePath = modelPath.erase(modelPath.find_last_of('/'), modelPath.length() - 1 - 1) + "/" + textureName; //Get texture path for loading through resource manager
-
-		HashedString pathString = HashedString(texturePath.c_str());
-
-		GetResourceManager()->AddResource<Texture>(pathString);
-		INFO("Loading texture at location {0}", texturePath);
-
-		textures.emplace_back(pathString.GetValue());
-		GetResourceManager()->GetResource<Texture>(pathString)->Unbind(0);
-
-		return textures;
-	}
-
-	void Model::LoadModel(tinygltf::Model& modelData)
-	{
-		//For now, I'm only looking at the first mesh. In the future, this should loop and create mesh objects. TODO!
-		auto& primitive = modelData.meshes[0].primitives[0];
-		std::vector<MeshData> meshData = LoadVertices(modelData, primitive);
-		std::vector<unsigned int> indices = LoadIndices(modelData, primitive);
-		m_Textures = LoadTexture(modelData, GetPath().GetString());
-		m_TotalIndices = static_cast<unsigned int>(indices.size());
-
-		VertexBuffer vertexBuffer;
-		vertexBuffer.SetLayout(
-			std::vector<BufferLayoutElement>{
-			BufferLayoutElement("a_Position", BufferVariableType::Float3, false), 
-			BufferLayoutElement("a_Normal", BufferVariableType::Float3, false),
-			BufferLayoutElement("a_TextureCoords", BufferVariableType::Float2, false) 
-			});
-		vertexBuffer.SetVertexData(meshData.data(), meshData.size() * sizeof(MeshData));
-
-		IndexBuffer indexBuffer;
-		indexBuffer.SetIndexData(indices);
-		m_Buffer.Initialize(vertexBuffer, indexBuffer);
 	}
 
 	void LAG::Model::Render(TransformComponent& transform, Entity* cameraEntity, Shader& shader, std::vector<std::pair<TransformComponent*, LightComponent*>>& lights)
@@ -228,11 +114,7 @@ namespace LAG
 		}
 		else shader.SetBool("a_UseLight", false);
 
-		
-		//Bind textures
-		for (size_t i = 0; i < m_Textures.size(); i++)
-			GetResourceManager()->GetResource<Texture>(m_Textures.at(i))->Bind(i);
-
-		m_Buffer.Render();
+		for (auto& it : m_Meshes)
+			it.Render();
 	}
 }
