+++ LAGGED/src/Platform/OpenGL/Renderer/GL_Model.h
@@ -1,32 +1,10 @@
 #pragma once
 #include "Core/Resources/Model.h"
-#include "Platform/Base/Renderer/Buffer.h"
 
-#include "glm/vec2.hpp"
-#include "glm/vec3.hpp"
-
-#include <map>
-#include <utility>
-#include <vector>
-
-namespace tinygltf
-{
-	class Model;
-	class Node;
-	struct Mesh;
-}
-
-typedef unsigned int GLuint;
+#include "GL_Mesh.h"
 
 namespace LAG
 {
-	struct MeshData
-	{
-		glm::vec3 vertices;
-		glm::vec3 normals;
-		glm::vec2 textureCoords;
-	};
-
 	class Entity;
 	class Model : public ModelBase
 	{
@@ -41,12 +19,6 @@ namespace LAG
 		bool Load() override;
 		bool Unload() override;
 
-		void LoadModel(tinygltf::Model& modelData) override;
-		
-		ArrayBuffer m_Buffer;
-
-		std::vector<size_t> m_Textures;
-
-		unsigned int m_TotalIndices = 0;
+		std::vector<Mesh> m_Meshes;
 	};
 }
